Move unsafe array operations into unsafe.ml
There have been 2 segfaults associated with the use of out of bounds
reads (which ended up reading and storing an invalid value on the OCaml
heap and causing a SIGSEGV later on during garbage collection).

Move all the unsafe array code to Unsafe.ml.
This makes it easier to replace them with safe versions for testing
purposes and find the sources of segfaults.

Eventually a special dune profile could be used to select a safe
implementation for testing purposes, but that is not done yet.

Also replace "%identity" in Stage.ml with Sys.opaque_identity (which is
always safe).

Signed-off-by: Edwin Török <edwin.torok@cloud.com>